--- app/event/forms.py
@@ -1,707 +1,0 @@
-from datetime import datetime, timedelta
-
-from django import forms
-from django.core.exceptions import ValidationError
-from django.utils import timezone
-
-from community.models import WEEKDAY_CHOICES, TAGS
-from user_account.forms import normalize_x_account
-from .datetime_lock import (
-    EVENT_DETAIL_DATETIME_LOCK_MESSAGE,
-    has_event_detail_duration_changed,
-    has_event_detail_start_time_changed,
-    is_event_detail_datetime_locked,
-)
-from .form_validators import validate_and_sanitize_pdf, validate_thumbnail_image
-from .models import EventDetail, RecurrenceRule, Event
-from .thumbnail import SLIDE_THUMBNAIL_ASPECT_RATIO_TEXT
-
-
-class EventSearchForm(forms.Form):
-    name = forms.CharField(
-        label='集会名',
-        required=False,
-        widget=forms.TextInput(attrs={
-            'class': 'form-control',
-            'placeholder': '集会名を入力',
-        })
-    )
-    weekday = forms.MultipleChoiceField(
-        label='曜日',
-        choices=WEEKDAY_CHOICES,
-        required=False,
-        widget=forms.CheckboxSelectMultiple()
-    )
-    tags = forms.MultipleChoiceField(
-        label='タグ',
-        choices=TAGS,
-        required=False,
-        widget=forms.CheckboxSelectMultiple()
-    )
-
-
-class EventCreateForm(forms.ModelForm):
-    class Meta:
-        model = Event
-        fields = ['date', 'start_time', 'duration']
-        widgets = {
-            'date': forms.DateInput(attrs={'type': 'date', 'class': 'form-control'}),
-            'start_time': forms.TimeInput(attrs={'type': 'time', 'class': 'form-control'}),
-            'duration': forms.NumberInput(attrs={'class': 'form-control'}),
-        }
-
-    def __init__(self, *args, **kwargs):
-        self.request = kwargs.pop('request', None)  # requestオブジェクトを受け取る
-        super().__init__(*args, **kwargs)
-        if self.request and self.request.user.is_authenticated:
-            membership = self.request.user.community_memberships.select_related('community').first()
-            if membership:
-                community = membership.community
-                self.fields['start_time'].initial = community.start_time  # Communityから初期値を設定
-                self.fields['duration'].initial = community.duration  # Communityから初期値を設定
-
-    def clean(self):
-        cleaned_data = super().clean()
-        # 過去日付のバリデーションを解除（何もしない）
-        return cleaned_data
-
-
-class RecurringEventForm(forms.Form):
-    """定期イベント作成フォーム"""
-    # 基本情報
-    base_date = forms.DateField(
-        label='開始日',
-        widget=forms.DateInput(attrs={
-            'type': 'date',
-            'class': 'form-control',
-            'id': 'id_base_date'
-        }),
-        help_text='定期イベントの最初の開催日'
-    )
-    
-    start_time = forms.TimeField(
-        label='開始時刻',
-        widget=forms.TimeInput(attrs={
-            'type': 'time',
-            'class': 'form-control',
-            'id': 'id_start_time'
-        }),
-        initial='22:00'
-    )
-    
-    duration = forms.IntegerField(
-        label='開催時間（分）',
-        initial=60,
-        min_value=15,
-        max_value=480,
-        widget=forms.NumberInput(attrs={
-            'class': 'form-control',
-            'id': 'id_duration'
-        })
-    )
-    
-    # 定期ルール
-    frequency = forms.ChoiceField(
-        label='開催パターン',
-        choices=RecurrenceRule.FREQUENCY_CHOICES,
-        initial='WEEKLY',
-        widget=forms.Select(attrs={
-            'class': 'form-control',
-            'id': 'id_frequency',
-            'onchange': 'updateFrequencyFields()'
-        })
-    )
-    
-    interval = forms.IntegerField(
-        label='間隔',
-        initial=1,
-        min_value=1,
-        max_value=12,
-        widget=forms.NumberInput(attrs={
-            'class': 'form-control',
-            'id': 'id_interval'
-        }),
-        help_text='何週間/月ごとに開催するか'
-    )
-    
-    week_of_month = forms.ChoiceField(
-        label='第N週',
-        required=False,
-        choices=[
-            (1, '第1'),
-            (2, '第2'),
-            (3, '第3'),
-            (4, '第4'),
-            (-1, '最終'),
-        ],
-        widget=forms.Select(attrs={
-            'class': 'form-control',
-            'id': 'id_week_of_month'
-        })
-    )
-    
-    custom_rule = forms.CharField(
-        label='カスタムルール',
-        required=False,
-        widget=forms.Textarea(attrs={
-            'class': 'form-control',
-            'id': 'id_custom_rule',
-            'rows': 3,
-            'placeholder': '例: 第2火曜日と第4金曜日\n例: 月末の平日\n例: 祝日を除く毎週月曜日',
-            'oninput': 'updatePreview()'
-        }),
-        help_text='複雑な開催パターンを日本語で記述してください'
-    )
-    
-    end_date = forms.DateField(
-        label='終了日',
-        required=False,
-        widget=forms.DateInput(attrs={
-            'type': 'date',
-            'class': 'form-control',
-            'id': 'id_end_date'
-        }),
-        help_text='定期イベントの終了日（省略可）'
-    )
-    
-    def __init__(self, *args, **kwargs):
-        self.request = kwargs.pop('request', None)
-        self.community = kwargs.pop('community', None)
-        super().__init__(*args, **kwargs)
-        
-        if self.community:
-            self.fields['start_time'].initial = self.community.start_time
-            self.fields['duration'].initial = self.community.duration
-    
-    def clean(self):
-        cleaned_data = super().clean()
-        frequency = cleaned_data.get('frequency')
-        
-        if frequency == 'MONTHLY_BY_WEEK' and not cleaned_data.get('week_of_month'):
-            raise ValidationError('第N曜日を選択してください')
-        
-        if frequency == 'OTHER' and not cleaned_data.get('custom_rule'):
-            raise ValidationError('カスタムルールを入力してください')
-        
-        base_date = cleaned_data.get('base_date')
-        end_date = cleaned_data.get('end_date')
-        
-        if base_date and end_date and end_date <= base_date:
-            raise ValidationError('終了日は開始日より後の日付を指定してください')
-        
-        return cleaned_data
-
-
-class EventDetailMediaFormMixin:
-    """スライドPDF・サムネ画像の検証と保存後処理を共通化するMixin。
-
-    EventDetailForm と LTApplicationEditForm の両方で同一実装が重複していたため抽出。
-    """
-
-    def clean_slide_file(self):
-        return validate_and_sanitize_pdf(self.cleaned_data.get('slide_file'))
-
-    def clean_thumbnail_image(self):
-        return validate_thumbnail_image(self.cleaned_data.get('thumbnail_image'))
-
-    def save(self, commit=True):
-        instance = super().save(commit=commit)
-        if commit:
-            from event.services.media_service import ensure_pdf_thumbnail
-            from twitter.signals import sync_slide_share_queue_image
-
-            ensure_pdf_thumbnail(instance, save=True)
-            sync_slide_share_queue_image(instance)
-        return instance
-
-
-class EventDetailForm(EventDetailMediaFormMixin, forms.ModelForm):
-    start_time = forms.TimeField(
-        label='開始時間',
-        widget=forms.TimeInput(
-            attrs={'type': 'time', 'class': 'form-control w-25'})  # Bootstrapのクラスを追加
-    )
-    duration = forms.IntegerField(
-        label='発表時間（分）',
-        min_value=1,
-        widget=forms.NumberInput(attrs={'class': 'form-control w-25'})
-    )
-    
-    # フィールドのラベルをオーバーライド
-    theme = forms.CharField(
-        label='テーマ',
-        required=False,  # 特別企画とブログでは不要なため
-        widget=forms.TextInput(attrs={'class': 'form-control'})
-    )
-    h1 = forms.CharField(
-        label='タイトル(H1)',
-        required=False,
-        widget=forms.TextInput(attrs={'class': 'form-control'})
-    )
-    contents = forms.CharField(
-        label='内容',
-        required=False,
-        widget=forms.Textarea(attrs={'class': 'form-control', 'rows': '8'})
-    )
-    
-    # 記事自動生成チェックボックス
-    generate_blog_article = forms.BooleanField(
-        label='記事を自動生成する',
-        required=False,
-        initial=True,  # デフォルトON
-        widget=forms.CheckboxInput(attrs={'class': 'form-check-input form-check-input-lg'}),
-        help_text='PDFまたはYouTube URLが設定されている場合、AIによって記事を自動生成します'
-    )
-
-    class Meta:
-        model = EventDetail
-        fields = ['detail_type', 'theme', 'speaker', 'start_time', 'duration', 'slide_file', 'slide_url',
-                  'thumbnail_image', 'youtube_url', 'h1', 'contents', 'generate_blog_article']
-        widgets = {
-            'detail_type': forms.RadioSelect(attrs={'class': 'form-check-input'}),
-            'youtube_url': forms.URLInput(attrs={'class': 'form-control'}),
-            'slide_url': forms.URLInput(attrs={'class': 'form-control'}),
-            'slide_file': forms.ClearableFileInput(attrs={'class': 'form-control-file', 'accept': '.pdf'}),
-            'thumbnail_image': forms.ClearableFileInput(attrs={'class': 'form-control-file', 'accept': 'image/*'}),
-            'speaker': forms.TextInput(attrs={
-                'class': 'form-control',
-                'placeholder': 'VRChat表示名を入力'
-            }),
-            'start_time': forms.TimeInput(attrs={'type': 'time', 'class': 'form-control w-25'}),
-            'duration': forms.NumberInput(attrs={'class': 'form-control w-25'}),
-            'theme': forms.TextInput(attrs={'class': 'form-control'}),
-            'h1': forms.TextInput(attrs={'class': 'form-control'}),
-            'contents': forms.Textarea(attrs={'class': 'form-control', 'rows': '8'}),
-        }
-        help_texts = {
-            'detail_type': '※ 記事の種類を選択してください。',
-            'contents': '※ Markdown形式で記述してください。',
-            'h1': '※ 空のときはテーマが使われます。',
-            'duration': '単位は分',
-            'youtube_url': 'YouTubeのURLの他、Discordのメッセージへのリンクも入力できます。',
-            'slide_file': '※ 記事生成に使うPDFです。まずここにスライドPDFをアップロードしてください（最大30MB）。',
-            'slide_url': '※ 任意。記事生成後、公開用の外部スライドURLがある場合に貼り付けてください。URL入力のみでは記事は生成されません。',
-            'thumbnail_image': (
-                f'※ 記事ページの上部に表示される画像です。'
-                f'アップロード時にスライドと同じ横長の比率（{SLIDE_THUMBNAIL_ASPECT_RATIO_TEXT}）へ自動トリミングします。'
-                'はみ出した部分は中央基準で切り取られます。'
-                '未設定でPDFがある場合はPDF保存時または記事生成時に自動設定されます。'
-            ),
-        }
-
-    # start_time と duration の初期値はEventCreateFormと同じにする
-    def __init__(self, *args, **kwargs):
-        self.request = kwargs.pop('request', None)
-        super().__init__(*args, **kwargs)
-        self.datetime_locked = False
-        self.datetime_lock_message = EVENT_DETAIL_DATETIME_LOCK_MESSAGE
-
-        if not self.instance.pk and self.request and self.request.user.is_authenticated:
-            membership = self.request.user.community_memberships.select_related('community').first()
-            if membership:
-                self.fields['start_time'].initial = membership.community.start_time
-            self.fields['duration'].initial = 30
-
-        if (
-            self.instance.pk
-            and self.request
-            and is_event_detail_datetime_locked(self.instance, self.request.user)
-        ):
-            self.datetime_locked = True
-            self.fields['start_time'].disabled = True
-            self.fields['duration'].disabled = True
-
-        # 既存の記事がある場合（更新時）は自動生成チェックボックスをOFFにする
-        if self.instance and self.instance.pk:
-            # meta_descriptionが存在する場合は記事が既に生成されていると判断
-            if self.instance.meta_description:
-                self.fields['generate_blog_article'].initial = False
-
-
-    def clean(self):
-        cleaned_data = super().clean()
-        detail_type = cleaned_data.get('detail_type')
-        is_datetime_locked = (
-            self.instance.pk
-            and self.request
-            and is_event_detail_datetime_locked(self.instance, self.request.user)
-        )
-
-        if is_datetime_locked:
-            if has_event_detail_start_time_changed(self.instance, self.data.get('start_time')):
-                self.add_error('start_time', EVENT_DETAIL_DATETIME_LOCK_MESSAGE)
-            if has_event_detail_duration_changed(self.instance, self.data.get('duration')):
-                self.add_error('duration', EVENT_DETAIL_DATETIME_LOCK_MESSAGE)
-
-        # 特別企画とブログの場合、非表示フィールドにデフォルト値を設定
-        if detail_type == 'SPECIAL':
-            # 特別企画のデフォルト値
-            cleaned_data['theme'] = 'Special Event'
-            cleaned_data['speaker'] = ''
-            # start_timeは入力されたものを使用（ただしVketロック中は元の値を維持）
-            if not getattr(self, 'datetime_locked', False):
-                cleaned_data['duration'] = 60
-        elif detail_type == 'BLOG':
-            # ブログのデフォルト値（h1があればthemeにコピー）
-            h1 = cleaned_data.get('h1', '')
-            cleaned_data['theme'] = h1 if h1 else 'Blog'
-            cleaned_data['speaker'] = ''
-            if not getattr(self, 'datetime_locked', False):
-                cleaned_data['start_time'] = self.instance.event.start_time if self.instance.pk else self.fields['start_time'].initial
-                cleaned_data['duration'] = 30
-
-        if is_datetime_locked:
-            cleaned_data['start_time'] = self.instance.start_time
-            cleaned_data['duration'] = self.instance.duration
-
-        return cleaned_data
-
-
-class LTApplicationEditForm(EventDetailMediaFormMixin, forms.ModelForm):
-    """LT申請者が自分の申請内容を編集するフォーム"""
-
-    generate_blog_article = forms.BooleanField(
-        label='記事を自動生成する',
-        required=False,
-        initial=True,
-        widget=forms.CheckboxInput(attrs={'class': 'form-check-input form-check-input-lg'}),
-        help_text='PDFまたはYouTube URLが設定されている場合、AIによって記事を自動生成します'
-    )
-
-    class Meta:
-        model = EventDetail
-        fields = ['theme', 'speaker', 'slide_file', 'slide_url', 'thumbnail_image', 'youtube_url', 'h1', 'contents',
-                  'generate_blog_article']
-        widgets = {
-            'theme': forms.TextInput(attrs={'class': 'form-control'}),
-            'speaker': forms.TextInput(attrs={
-                'class': 'form-control',
-                'placeholder': 'VRChat表示名を入力'
-            }),
-            'slide_url': forms.URLInput(attrs={'class': 'form-control'}),
-            'slide_file': forms.ClearableFileInput(attrs={'class': 'form-control-file', 'accept': '.pdf'}),
-            'thumbnail_image': forms.ClearableFileInput(attrs={'class': 'form-control-file', 'accept': 'image/*'}),
-            'youtube_url': forms.URLInput(attrs={'class': 'form-control'}),
-            'h1': forms.TextInput(attrs={'class': 'form-control'}),
-            'contents': forms.Textarea(attrs={'class': 'form-control', 'rows': '8'}),
-        }
-        help_texts = {
-            'contents': '※ Markdown形式で記述してください。',
-            'h1': '※ 空のときはテーマが使われます。',
-            'youtube_url': 'YouTubeのURLの他、Discordのメッセージへのリンクも入力できます。',
-            'slide_file': '※ 記事生成に使うPDFです。まずここにスライドPDFをアップロードしてください（最大30MB）。',
-            'slide_url': '※ 任意。記事生成後、公開用の外部スライドURLがある場合に貼り付けてください。URL入力のみでは記事は生成されません。',
-            'thumbnail_image': (
-                f'※ 記事ページの上部に表示される画像です。'
-                f'アップロード時にスライドと同じ横長の比率（{SLIDE_THUMBNAIL_ASPECT_RATIO_TEXT}）へ自動トリミングします。'
-                'はみ出した部分は中央基準で切り取られます。'
-                '未設定でPDFがある場合はPDF保存時または記事生成時に自動設定されます。'
-            ),
-        }
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        # 記事未生成ならON、生成済みならOFF
-        has_article = self.instance and self.instance.pk and self.instance.meta_description
-        self.initial['generate_blog_article'] = not has_article
-
-
-RECURRENCE_CHOICES = [
-    ('none', '単発イベント'),
-    ('weekly', '毎週'),
-    ('biweekly', '隔週'),
-    ('monthly_by_day', '毎月同じ曜日（例：第2月曜日）'),
-    ('monthly_by_date', '毎月同じ日（例：毎月15日）'),
-]
-
-CALENDAR_WEEKDAY_CHOICES = [
-    ('MO', '月曜日'),
-    ('TU', '火曜日'),
-    ('WE', '水曜日'),
-    ('TH', '木曜日'),
-    ('FR', '金曜日'),
-    ('SA', '土曜日'),
-    ('SU', '日曜日'),
-]
-
-WEEK_NUMBER_CHOICES = [
-    (1, '第1'),
-    (2, '第2'),
-    (3, '第3'),
-    (4, '第4'),
-    (-1, '最終'),
-]
-
-
-class GoogleCalendarEventForm(forms.Form):
-
-    community_name = forms.CharField(
-        label='コミュニティ',
-        required=False,
-        widget=forms.TextInput(attrs={'class': 'form-control', 'readonly': True}),
-    )
-
-    start_date = forms.DateField(
-        label='開始日',
-        widget=forms.DateInput(attrs={
-            'type': 'date',
-            'class': 'form-control'
-        }),
-        help_text='イベントの開始日を選択してください'
-    )
-
-    start_time = forms.TimeField(
-        label='開始時刻',
-        widget=forms.TimeInput(attrs={
-            'type': 'time',
-            'class': 'form-control'
-        }),
-        initial='22:00',
-        help_text='イベントの開始時刻を選択してください'
-    )
-
-    duration = forms.IntegerField(
-        label='開催時間（分）',
-        initial=60,
-        min_value=15,
-        max_value=480,
-        widget=forms.NumberInput(attrs={'class': 'form-control'}),
-        help_text='イベントの開催時間を分単位で入力してください'
-    )
-
-    recurrence_type = forms.ChoiceField(
-        choices=RECURRENCE_CHOICES,
-        label='開催周期',
-        initial='none',
-        widget=forms.RadioSelect(attrs={'class': 'form-check-input'}),
-        help_text='イベントの開催周期を選択してください'
-    )
-
-    weekday = forms.ChoiceField(
-        choices=CALENDAR_WEEKDAY_CHOICES,
-        label='曜日',
-        required=False,
-        widget=forms.Select(attrs={'class': 'form-control'}),
-        help_text='週次・隔週・毎月同じ曜日の場合、開催する曜日を選択してください'
-    )
-
-    monthly_day = forms.IntegerField(
-        label='開催日',
-        required=False,
-        min_value=1,
-        max_value=31,
-        widget=forms.NumberInput(attrs={'class': 'form-control'}),
-        help_text='毎月同じ日に開催する場合、その日付を選択してください（1-31）'
-    )
-
-    week_number = forms.ChoiceField(
-        choices=WEEK_NUMBER_CHOICES,
-        label='週',
-        required=False,
-        widget=forms.Select(attrs={'class': 'form-control'}),
-        help_text='毎月同じ曜日の場合、第何週かを選択してください'
-    )
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        # コミュニティが選択されている場合、そのコミュニティの設定を初期値として使用
-        if 'initial' in kwargs and 'community' in kwargs['initial']:
-            community = kwargs['initial']['community']
-            if community:
-                self.fields['start_time'].initial = community.start_time
-                self.fields['duration'].initial = community.duration
-
-    def clean(self):
-        cleaned_data = super().clean()
-        start_date = cleaned_data.get('start_date')
-        recurrence_type = cleaned_data.get('recurrence_type')
-
-        if start_date:
-            # 過去の日付バリデーションを解除
-            pass
-
-            # 6ヶ月以上先の日付をチェック
-            max_date = datetime.now().date() + timedelta(days=180)
-            if start_date > max_date:
-                raise ValidationError('6ヶ月以上先の日付は選択できません')
-
-        # 定期開催の場合の追加バリデーション
-        if recurrence_type in ['weekly', 'biweekly']:
-            if not cleaned_data.get('weekday'):
-                raise ValidationError('週次・隔週の場合は曜日を選択してください')
-
-        elif recurrence_type == 'monthly_by_day':
-            if not cleaned_data.get('weekday'):
-                raise ValidationError('毎月同じ曜日の場合は曜日を選択してください')
-            if not cleaned_data.get('week_number'):
-                raise ValidationError('毎月同じ曜日の場合は第何週かを選択してください')
-
-        elif recurrence_type == 'monthly_by_date':
-            if not cleaned_data.get('monthly_day'):
-                raise ValidationError('月次（日付指定）の場合は開催日を選択してください')
-
-            # 31日がない月もあるため、28日までを推奨
-            if cleaned_data.get('monthly_day') > 28:
-                self.add_error(
-                    'monthly_day', '月末の日付は月によって異なるため、28日以前の選択を推奨します')
-
-        return cleaned_data
-
-
-# LT申請時のデフォルト発表時間（分）
-DEFAULT_LT_DURATION = 15
-
-
-class LTApplicationForm(forms.Form):
-    """LT発表の申請フォーム"""
-
-    event = forms.ModelChoiceField(
-        queryset=Event.objects.none(),
-        label='開催日',
-        widget=forms.Select(attrs={'class': 'form-control'}),
-        help_text='発表したい開催日を選択してください'
-    )
-
-    theme = forms.CharField(
-        label='テーマ',
-        max_length=100,
-        widget=forms.TextInput(attrs={
-            'class': 'form-control',
-            'placeholder': '発表テーマを入力してください'
-        }),
-        help_text='発表内容を簡潔に表すテーマを入力してください'
-    )
-
-    speaker = forms.CharField(
-        label='発表者名',
-        max_length=200,
-        widget=forms.TextInput(attrs={
-            'class': 'form-control',
-            'placeholder': 'VRChat表示名を入力してください'
-        }),
-        help_text='VRChatの表示名を入力してください。送信するとアカウントの表示名としても保存されます。'
-    )
-
-    duration = forms.IntegerField(
-        label='発表時間（分）',
-        initial=DEFAULT_LT_DURATION,
-        min_value=5,
-        max_value=60,
-        widget=forms.NumberInput(attrs={'class': 'form-control'}),
-        help_text='希望する発表時間を分単位で入力してください（5〜60分）'
-    )
-
-    x_account = forms.CharField(
-        label='X (Twitter) アカウント',
-        max_length=64,
-        required=False,
-        widget=forms.TextInput(attrs={
-            'class': 'form-control',
-            'placeholder': 'noricha_vr',
-            'autocomplete': 'off',
-        }),
-        help_text='任意。@ や https://x.com/ のURLでもOK。アカウント情報にも保存されます。'
-    )
-
-    additional_info = forms.CharField(
-        label='追加情報',
-        required=False,
-        widget=forms.Textarea(attrs={
-            'class': 'form-control',
-            'rows': 6,
-        }),
-        help_text='主催者が設定したテンプレートに沿って入力してください'
-    )
-
-    def __init__(self, *args, **kwargs):
-        self.community = kwargs.pop('community', None)
-        self.user = kwargs.pop('user', None)
-        super().__init__(*args, **kwargs)
-
-        if self.community:
-            # コミュニティのデフォルトLT発表時間を設定
-            self.fields['duration'].initial = self.community.default_lt_duration
-
-            # accepts_lt_application=True かつ未来のイベントのみ
-            today = timezone.now().date()
-            self.fields['event'].queryset = Event.objects.filter(
-                community=self.community,
-                accepts_lt_application=True,
-                date__gte=today
-            ).order_by('date', 'start_time')
-
-            # テンプレートが設定されている場合は初期値として編集可能にする
-            if self.community.lt_application_template:
-                self.fields['additional_info'].initial = self.community.lt_application_template
-            else:
-                self.fields['additional_info'].help_text = (
-                    '追加で伝えたい情報があれば入力してください'
-                )
-
-        if self.user and self.user.is_authenticated:
-            self.fields['speaker'].initial = self.user.display_label
-            self.fields['x_account'].initial = self.user.x_account
-
-    def clean_speaker(self):
-        """発表者名を EventDetail.speaker と user.display_name 用に検証する。"""
-        speaker = (self.cleaned_data.get('speaker') or '').strip()
-        if not speaker:
-            raise ValidationError('発表者名を入力してください')
-        if len(speaker) > EventDetail._meta.get_field('speaker').max_length:
-            raise ValidationError('発表者名は200文字以下で入力してください')
-        return speaker
-
-    def clean_x_account(self):
-        return normalize_x_account(self.cleaned_data.get('x_account', ''))
-
-    def clean_additional_info(self):
-        """追加情報のバリデーション"""
-        additional_info = self.cleaned_data.get('additional_info', '')
-
-        if not self.community or not self.community.lt_application_template:
-            return additional_info
-
-        template = self.community.lt_application_template
-
-        # テンプレートと同一内容のチェック
-        if additional_info.strip() == template.strip():
-            raise ValidationError('テンプレートの各項目を入力してください')
-
-        return additional_info
-
-
-class LTApplicationReviewForm(forms.Form):
-    """LT申請の承認/却下フォーム"""
-
-    ACTION_CHOICES = [
-        ('approve', '承認する'),
-        ('reject', '却下する'),
-    ]
-
-    action = forms.ChoiceField(
-        choices=ACTION_CHOICES,
-        label='アクション',
-        widget=forms.RadioSelect(attrs={'class': 'form-check-input'}),
-        initial='approve'
-    )
-
-    rejection_reason = forms.CharField(
-        label='却下理由',
-        required=False,
-        widget=forms.Textarea(attrs={
-            'class': 'form-control',
-            'rows': 3,
-            'placeholder': '却下する場合は理由を入力してください'
-        }),
-        help_text='却下する場合は、申請者に伝える理由を入力してください'
-    )
-
-    def clean(self):
-        cleaned_data = super().clean()
-        action = cleaned_data.get('action')
-        rejection_reason = cleaned_data.get('rejection_reason')
-
-        if action == 'reject' and not rejection_reason:
-            raise ValidationError('却下する場合は理由を入力してください')
-
-        return cleaned_data
