+++ app/event/forms/__init__.py
@@ -1,0 +1,20 @@
+"""event.forms パッケージ互換 re-export
+
+既存の `from event.forms import EventDetailForm` 等を後方互換で動作させる。
+"""
+from event.forms.mixins import EventDetailMediaFormMixin  # noqa: F401
+from event.forms.event import EventSearchForm, EventCreateForm  # noqa: F401
+from event.forms.recurrence import (  # noqa: F401
+    RecurringEventForm,
+    RECURRENCE_CHOICES,
+    CALENDAR_WEEKDAY_CHOICES,
+    WEEK_NUMBER_CHOICES,
+)
+from event.forms.event_detail import EventDetailForm  # noqa: F401
+from event.forms.lt_application import (  # noqa: F401
+    DEFAULT_LT_DURATION,
+    LTApplicationEditForm,
+    LTApplicationForm,
+    LTApplicationReviewForm,
+)
+from event.forms.calendar import GoogleCalendarEventForm  # noqa: F401
